tools/transcribe-batch/pyproject.toml
[project]
name = "sz-call-analytics"
version = "0.1.0"
description = "Folder-based WAV transcription with Azure AI Foundry."
requires-python = ">=3.11"
dependencies = [
    "openai",
    "python-dotenv",
]

[tool.uv]
package = false
